+++ src/pages/Facility/services/inventory/externalSupply/requestOrder/RequestOrderShow.tsx
@@ -269,15 +269,14 @@ export function RequestOrderShow({
             </div>
           </div>
           <div className="flex items-center justify-end gap-2">
-            {isRequester &&
-              requestOrder.status !== RequestOrderStatus.completed && (
-                <Button variant="outline" asChild>
-                  <Link href={`${requestOrderId}/edit`}>
-                    <Edit /> {t("edit")}
-                    <ShortcutBadge actionId="edit-order" />
-                  </Link>
-                </Button>
-              )}
+            {isRequester && (
+              <Button variant="outline" asChild>
+                <Link href={`${requestOrderId}/edit`}>
+                  <Edit /> {t("edit")}
+                  <ShortcutBadge actionId="edit-order" />
+                </Link>
+              </Button>
+            )}
 
             {canAddSupplyRequests && (
               <Button
